python/ipxl/ipxlPython/requirements.txt
bleak
crcmod
Pillow
